move existing App to Form component
