+++ src/js/components/Nav.js
@@ -5,7 +5,6 @@ import {
   Toolbar,
   IconButton,
   Button,
-  Link,
   Grid,
 } from "@material-ui/core";
 import MenuIcon from "@material-ui/icons/Menu";
